refactor: shared SupportDialog, remove Settings donation card
- Extract SupportDialog into shared component used by Header + HomePage
- HomePage milestone button now opens dialog instead of external link
- Remove donation card from SettingsPage
- Rename "Buy us a coffee" → "Buy the developer a coffee"
