+++ src/components/layout/Header.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom'
 import { Menu, X, LogOut, ChevronDown, Coffee, Search, BookOpen } from 'lucide-react'
 import { track } from '@vercel/analytics'
 import { Button } from '@/components/ui/button'
-import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogDescription } from '@/components/ui/dialog'
+import { SupportDialog } from './SupportDialog'
 import { AVAILABLE_MODULES, DEFAULT_MODULE, SETTINGS_KEYS } from '@/lib/constants'
 import { useTranslation } from '@/lib/i18n'
 import type { User } from '@supabase/supabase-js'
@@ -14,9 +14,6 @@ interface HeaderProps {
   onToggleMenu: () => void
   onSignOut: () => void
 }
-
-// Hardcoded — not in i18n translations to reduce exposure if site is compromised
-const MOBILEPAY_NUMBER = ['+45', '5272', '8520'].join(' ')
 
 export function Header({ user, menuOpen, onToggleMenu, onSignOut }: HeaderProps) {
   const { locale, setLocale, t } = useTranslation()
@@ -148,23 +145,7 @@ export function Header({ user, menuOpen, onToggleMenu, onSignOut }: HeaderProps)
         )}
       </div>
 
-      <Dialog open={supportOpen} onOpenChange={setSupportOpen}>
-        <DialogContent className="max-w-sm">
-          <DialogHeader>
-            <DialogTitle className="flex items-center gap-2">
-              <Coffee className="h-5 w-5 text-pink-500" />
-              {t('support.dialogTitle')}
-            </DialogTitle>
-            <DialogDescription>{t('support.dialogDesc')}</DialogDescription>
-          </DialogHeader>
-          <div className="flex flex-col items-center gap-3 py-2">
-            <div className="rounded-lg border-2 border-pink-200 dark:border-pink-800 bg-pink-50 dark:bg-pink-950 px-6 py-3">
-              <p className="text-2xl font-bold tracking-wider text-center select-all">{MOBILEPAY_NUMBER}</p>
-            </div>
-            <p className="text-xs text-muted-foreground text-center">{t('support.thankYou')}</p>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <SupportDialog open={supportOpen} onOpenChange={setSupportOpen} />
     </header>
   )
 }
